+++ apps/gateway/src/routes/api/v3/projects/versions/documents/destroy/destroy.handler.ts
@@ -4,14 +4,11 @@ import {
 } from '@latitude-data/core/repositories'
 import { findProjectById } from '@latitude-data/core/queries/projects/findById'
 import { destroyDocument } from '@latitude-data/core/services/documents/destroyDocument'
-import { destroyDocumentUnsafe } from '@latitude-data/core/services/documents/destroyDocumentUnsafe'
-import { BadRequestError, NotFoundError } from '@latitude-data/core/lib/errors'
+import { NotFoundError } from '@latitude-data/core/lib/errors'
 
 import type { DestroyDocumentRoute } from './destroy.route'
 import { AppRouteHandler } from '$/openApi/types'
 import { HEAD_COMMIT } from '@latitude-data/constants'
-import { assertCanEditCommit } from '@latitude-data/core/lib/assertCanEditCommit'
-import { Result } from '@latitude-data/core/lib/Result'
 
 // @ts-expect-error: Types are not working as expected
 export const destroyDocumentHandler: AppRouteHandler<
@@ -19,7 +16,6 @@ export const destroyDocumentHandler: AppRouteHandler<
 > = async (c) => {
   const workspace = c.get('workspace')
   const { projectId, versionUuid, documentPath } = c.req.valid('param')
-  const { force } = c.req.valid('query')
 
   const commitsScope = new CommitsRepository(workspace.id)
   const docsScope = new DocumentVersionsRepository(workspace.id)
@@ -39,14 +35,6 @@ export const destroyDocumentHandler: AppRouteHandler<
     })
     .then((r) => r.unwrap())
 
-  const canEditCommit = await assertCanEditCommit(commit)
-
-  if (!Result.isOk(canEditCommit) && !force) {
-    throw new BadRequestError(
-      'Cannot modify a merged commit. Use force=true to allow deletions on the live commit.',
-    )
-  }
-
   const docResult = await docsScope.getDocumentByPath({
     commit,
     path: documentPath,
@@ -56,15 +44,9 @@ export const destroyDocumentHandler: AppRouteHandler<
   }
   const document = docResult.unwrap()
 
-  if (Result.isOk(canEditCommit)) {
-    await destroyDocument({ document, commit, workspace }).then((r) =>
-      r.unwrap(),
-    )
-  } else {
-    await destroyDocumentUnsafe({ document, commit, workspace }).then((r) =>
-      r.unwrap(),
-    )
-  }
+  await destroyDocument({ document, commit, workspace }).then((r) =>
+    r.unwrap(),
+  )
 
   return c.json(
     {